+++ NEWS.md
@@ -1,3 +1,4 @@
 # pipfun (development version)
 
+* Add `save_to_gh` function
 * add checks for active package `check_pkg_active()`